Added AFSettingsXMLConfigurationProcessor
