+++ kie-wb-common-services/kie-wb-common-services-backend/src/test/settings.xml
@@ -1,0 +1,45 @@
+<settings xmlns="http://maven.apache.org/SETTINGS/1.0.0"
+          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+          xsi:schemaLocation="http://maven.apache.org/SETTINGS/1.0.0
+                      http://maven.apache.org/xsd/settings-1.0.0.xsd">
+  <localRepository/>
+  <interactiveMode/>
+  <usePluginRegistry/>
+  <offline/>
+  <pluginGroups/>
+  <servers/>
+  <mirrors/>
+  <proxies/>
+  <profiles>
+    <profile>
+      <id>additional-repositories</id>
+      <repositories>
+        <repository>
+          <id>kie-group</id>
+          <url>https://repository.jboss.org/nexus/content/groups/kie-group/</url>
+        </repository>
+        <repository>
+          <id>kie-internal-group</id>
+          <url>https://repository.jboss.org/nexus/content/groups/kie-internal-group/</url>
+        </repository>
+
+      </repositories>
+      <pluginRepositories>
+        <pluginRepository>
+          <id>kie-group</id>
+          <url>https://repository.jboss.org/nexus/content/groups/kie-group/</url>
+        </pluginRepository>
+        <pluginRepository>
+          <id>kie-internal-group</id>
+          <url>https://repository.jboss.org/nexus/content/groups/kie-internal-group/</url>
+        </pluginRepository>
+
+      </pluginRepositories>
+    </profile>
+  </profiles>
+
+  <activeProfiles>
+    <activeProfile>additional-repositories</activeProfile>
+  </activeProfiles>
+
+</settings>
